Added veritcal/horizontal table line params
